Log training_wrapper's output to file
The log directory is defined in settings.py. In the future, an optional argument will allow redirecting the log file elsewhere.

Additionally, current working directory when executing training is set to the project's root folder. This allows us to use project relative paths more easily, which are easier to understand, adjust, and parse.