Improve existing argument override checks
Namespace attributes will receive a value of 'None' if the argument is not present. Rather than setting a default value, we check arguments against a None value before overriding the configuration. An argument can still contain an empty string (this is not equivalent to a None value) which allows us to purposefully set empty arguments from the command line. An example of this is overriding the '--env' argument with an empty string. This would prevent mlagents-learn from looking for a built environment.

A side effect / benefit from this change is the '--timestamp' argument now requires either a true or false value when it is present. Likewise, this allows us to purposefully disable a timestamp if we want to use the --load argument to resume training.