__main__ will now determine and set the run_id time-stamp
training_wrapper was previously responsible for saving to history the training commands it executed on mlagents-learn, as it had authority over the run_id time-stamp. This caused the optional brain export path to be lost from history as the argument was consumed by training_wrapper and not passed on to mlagents-learn (and not saved in history). This resulted in brains not being exported by training_wrapper when using the --resume argument on __main__.

Shifting the time-stamp determination into __main__ resolves this issue.

Side effects of this change are that log filenames will include a time-stamp (which is a regression) and training_wrapper commands are saved in the history, rather than mlagent-learn commands.