Add configuration setting for training run log filename
Log files are generated by training_wrapper using the run-id. This results in a large number of unique logs being generated as a result of the time-stamp being included. The addition of a log filename configuration allows for all of these unique log files to be more usefully condensed into one. If a logfile name isn't configured, it defaults to the time-stamp free version of the run-id.