Simplify test_override_configuration_values()
As this test should be operating directly on override_configuration_values() and not create_command(), we now set up parameters for calling override_configuration_values() directly and check the mutated TrainingCommand object.

This is unfortunately still not a solid unit test, as it relies on method calls into TrainingCommand, but I'm not going to setup mock objects for all of them just yet.