Module execution calls now use absolute rather than relative paths
The desire to use the MLAgents project root folder as the base for all other relative paths in the project is mostly driven by the potential disconnect between the current working directory at execution (which could be any folder) and the relative path defined in a grimagents configuration file. I think it will be better to take a hands-off approach and allow either relative or absolute paths to be used for the 'trainer-config' configuration value and encourage grimagents execution from the MLAgents project folder in the readme.

This commit begins the process of removing process calls that force the current working directory to a specific folder.