Set 'buffer_size' based on 'buffer_size_multiple', if present
As 'buffer_size' should always be a multiple of the 'batch_size', this makes it impossible to perform a grid search on one or the other using static values. A special 'buffer_size_multiple' value allows GridSearch to dynamically set the 'buffer_size' based directly on the 'batch_size' and still meaningfully search.