Implement Bayesian optimization loading
Loads Bayesian optimization observations from a json file. Importantly, the save operation does NOT append to the observations file, but overwrites it completely. The load operation is performed first to prevent a FileNotFound exception, but this means that many successive save and load operations will not remember ancestral observations.

The BayesianOptimization load_logs() method accepts a list of logs, however there's no unwieldy way to support this with command line arguments directly. A potential solution is instead of accepting a file for both arguments, a folder is used instead.