Remove '--new-window' argument from grimagents
Continued support for this argument is the main obstacle in adding Linux support to grimagents as there doesn't appear to be one universal way to run commands in a new terminal under Linux.

This functionality was required when grim configuration files could specify multiple trainer config files to execute training runs on in parallel. Loading multiple grim configuration files was replaced with grimsearch and new-window functionality is now largely superficial. The user can always open a terminal window manually to run new commands in.