Fix issue where test was not being run on a SearchCommand method
The unit test targeted at perform_search_with_configuration() was not being executed on the actual method because the patch_search_command() fixture was patching it out. This extracts patch_perform_search_with_configuration() into its own fixture and included in all tests that require it.

Deciding which methods to include in any given patch has caused some difficulty. I think moving forward the best way to approach it is that patches that operate on a whole class should only exclude external module method calls, and any internal methods that need to be patched should either be applied within individual tests, or a separate patch should be made for that specific method.