Use the correct dictionary for Bayesian search 'best configuration'
Previously, the entire 'optimizer_max' structure was being sent to the sanitize_paramter_values() method, which is a dictionary of dictionaries. We desire only the 'params' value, which is a dictionary of parameter names and values.

Additionally, this fixes an issue where the best configuration for Bayesian searches were not being displayed or saved correctly.