Remove test workflow from the Windows platform
The test workflow currently fails on the windows platform test runner when pipenv attempts to set up a virtual environment. The lack of useful debugging information regarding the failure is making this difficult to get to the bottom of. I currently suspect its something to do with pipenv's console output (maybe something to do with its install emojii spam) or something about the test runner environments that does not like pipenv.

This is not ideal, but the Linux and OSX tests will catch major package issues and I should see windows specific issues pop up in my local development environment.