Update example configuration file to match readme
Pointing the grimagents configuration file to the package's 'etc' folder instead of the 'config' folder included with MLAgents projects is much less confusing and makes the configuration more self-contained.

This should also make it quicker for a new user of grimagents to get it up and running. All that needs doing is a training build of 3DBall be made and placed in the MLAgents 'build' folder.